requirements.txt
flask==3.0.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0